backend/services/user-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
google-auth==2.25.2
requests==2.31.0

